requirements.txt
pytest==9.0.2
pytest-bdd==8.1.0
appium-python-client==5.2.7
PyYAML==6.0.3